requirements.txt
google-adk==1.14.0
yfinance==0.2.54
mcp[cli]==1.9.0
fastmcp==2.3.3
python-dotenv==1.1.0
google-cloud-logging
